feat(structure): renaming of generic pane components
